fix db units

+++ server/memphis_handlers_monitoring.go
@@ -1489,9 +1489,9 @@ func getDbStorageSize() (int64, int64, error) {
 		return 0, 0, err
 	}
 
-	dbStorageSize := sbStats.Lookup("dataSize").Int64() + sbStats.Lookup("indexSize").Int64()
-	totalSize := sbStats.Lookup("fsTotalSize").Int64()
-	return dbStorageSize, totalSize, nil
+	dbStorageSize := sbStats.Lookup("dataSize").Double() + sbStats.Lookup("indexSize").Double()
+	totalSize := sbStats.Lookup("fsTotalSize").Double()
+	return int64(dbStorageSize), int64(totalSize), nil
 }
 
 func getUnixStorageSize() (int64, error) {
